+++ packages/cli/src/lib/spinner.ts
@@ -1,4 +1,3 @@
-// import type { FunctionType } from '@codeleap/common'
 import { createSpinner, Options } from 'nanospinner'
 
 type SpinWhileNotCompletedOpts = Options & {