+++ packages/mobile/src/components/Image/index.tsx
@@ -6,6 +6,8 @@ import {
   arePropsEqual,
   FormTypes,
   TypeGuards,
+  useBooleanToggle,
+  getNestedStylesByKey,
 } from '@codeleap/common'
 import { ComponentPropsWithoutRef } from 'react'
 import {
@@ -18,16 +20,20 @@ import {
 } from 'react-native'
 import { FastImage } from '../../modules/fastImage'
 import {
+  ImageComposition,
   ImageStyles,
 } from './styles'
 import { useImageSpotlight } from '../ImageView/Spotlight'
 import { Touchable } from '../Touchable'
 import { isFile, toMultipartFile } from '../../utils'
+import { LoadingOverlay, LoadingOverlayProps } from '../LoadingOverlay'
+import { StylesOf } from '../../types'
 export * from './styles'
 type NativeImageProps = ComponentPropsWithoutRef<typeof NativeImage>
 export type ImageProps = Omit<NativeImageProps, 'source' | 'style'> & {
   variants?: ComponentVariants<typeof ImageStyles>['variants']
   fast?: boolean
+  styles?: StylesOf<ImageComposition>
   style?: StyleProp<ImageStyle | TextStyle | ViewStyle>
   source:
     | (NativeImageProps['source'] & {
@@ -38,24 +44,32 @@ export type ImageProps = Omit<NativeImageProps, 'source' | 'style'> & {
     | number
   resizeMode?: keyof typeof FastImage.resizeMode
   spotlight?: string
-  maintainAspectRatioFrom?: 'width' | 'height' | 'none'
+  maintainAspectRatio?: boolean
+  withLoadingOverlay?: boolean | React.FC<LoadingOverlayProps>
 }
+
+
+
 
 export const ImageComponent: React.FC<ImageProps> = (props) => {
   const {
     variants,
     style,
+    styles: componentStyleSheet = {},
     fast = true,
     spotlight = null,
     resizeMode = 'contain',
     source,
-    maintainAspectRatioFrom = 'height',
+    withLoadingOverlay = false,
+    maintainAspectRatio = true,
     ...imageProps
   } = props
 
-  const variantStyles = useDefaultComponentStyle<'u:Image', typeof ImageStyles>('u:Image', { variants })
+  const variantStyles = useDefaultComponentStyle<'u:Image', typeof ImageStyles>('u:Image', { variants, styles: componentStyleSheet,transform: StyleSheet.flatten })
+  const [loading, setLoading] = React.useState(false)
 
   const styles = StyleSheet.flatten([variantStyles.wrapper, style])
+  
   let imSource = source
   if (isFile(imSource)) {
     imSource = toMultipartFile(imSource)
@@ -72,10 +86,11 @@ export const ImageComponent: React.FC<ImageProps> = (props) => {
   }
 
   const aspectRatioStyle = React.useMemo(() => {
-    if (maintainAspectRatioFrom === 'none' || !imSource) return null
+    if (!maintainAspectRatio|| !imSource) return null
     try {
       // @ts-ignore
       const assetSource = NativeImage.resolveAssetSource(imSource)
+
       const aspectRatio = assetSource.width / assetSource.height
 
       if (Number.isNaN(aspectRatio)) {
@@ -88,8 +103,38 @@ export const ImageComponent: React.FC<ImageProps> = (props) => {
       return null
     }
 
-  }, [maintainAspectRatioFrom, imSource])
+  }, [maintainAspectRatio, imSource])
 
+  const loadEndedEarly = React.useRef(false)
+
+  const loadProps = React.useRef({
+    onLoadStart: () => {
+      if(withLoadingOverlay) {
+        setTimeout(() => {
+          if(!loadEndedEarly.current){
+            setLoading(true)
+          }
+        },60)
+      }
+    }, onLoadEnd: () => {
+      loadEndedEarly.current = true
+      if(withLoadingOverlay) setLoading(false)
+    }
+  })
+
+  const Loading = TypeGuards.isFunction(withLoadingOverlay) ? withLoadingOverlay : LoadingOverlay
+  const showLoading = !!withLoadingOverlay
+
+  const overlayStyle = React.useMemo(() => getNestedStylesByKey('overlay', variantStyles), [variantStyles])
+
+  const loadingElement = React.useMemo(() =>  {
+    return showLoading ? (
+      <Loading visible={loading} styles={overlayStyle}/>
+  
+      ) : null
+  }, [showLoading, loading])
+    
+  
   if (fast) {
     return (
       <Wrapper {...wrapperProps}>
@@ -100,14 +145,23 @@ export const ImageComponent: React.FC<ImageProps> = (props) => {
           tintColor={styles?.tintColor}
           source={imSource}
           resizeMode={FastImage.resizeMode[resizeMode || 'contain']}
+          {...loadProps.current}
           {...imageProps}
         />
+        {loadingElement}
       </Wrapper>
     )
   }
   return <Wrapper {...wrapperProps}>
-    <NativeImage style={[aspectRatioStyle, styles]} resizeMode={resizeMode} source={imSource} {...(imageProps as any)} />
-
+    <NativeImage 
+      style={[aspectRatioStyle, styles]} 
+      resizeMode={resizeMode} 
+      source={imSource} {...(imageProps as any)}  
+      {...loadProps.current}
+    />
+    {
+      loadingElement
+    }
   </Wrapper>
 }