+++ packages/mobile/src/components/Image/styles.ts
@@ -1,6 +1,7 @@
 import { createDefaultVariantFactory, includePresets } from '@codeleap/common'
+import { LoadingOverlayComposition } from '../LoadingOverlay/styles'
 
-export type ImageComposition = 'wrapper' | 'touchable'
+export type ImageComposition = 'wrapper' | 'touchable' | `overlay${Capitalize<LoadingOverlayComposition>}`
 const createImageStyle = createDefaultVariantFactory<ImageComposition>()
 
 const presets = includePresets((styles) => createImageStyle(() => ({ wrapper: styles, touchable: styles })),