+++ packages/mobile/src/components/components.ts
@@ -32,11 +32,9 @@ export * from './MultiSelect'
 export * from './AutoComplete'
 export * from './Grid'
 export * from './InputLabel'
+export * from './LoadingOverlay'
 export * from './ActionIcon'
 
-import * as Legacy from './legacy'
-export {
-  Legacy,
-}
+
 
 export * from './defaultStyles'